Update DVELP's Vision and Values
Why:

* DVELP is evolving and the time has come to review the original goals
and belief sets that were established at the outset and recalibrate them
in the context of where we are today and where we want to be.

This change addresses the need by:

* Updating the vision and values to reflect on the current values that I
believe represent us a team and to give structure and clarity on what is
important to us, for those who will join in the short and long-term.